+++ src/test/scala/com/github/pawelkrol/CPU6502/Operations/AbsoluteYSpec.scala
@@ -11,37 +11,37 @@ class AbsoluteYSpec extends FunOperationsSpec {
 
   describe("absolute,x addressing mode") {
     testOpCode(OpCode_ORA_ABSY) {
-      it { expect { operation }.toAdvancePC(0x03) }
-      it { expect { operation }.toUseCycles(0x04) }
+      it { expect2 { operation }.toAdvancePC(0x03) }
+      it { expect2 { operation }.toUseCycles(0x04) }
 
       context("AC = $00; YR = $02") { AC = 0x00; YR = 0x02; ZF = true; SF = false } {
         context("ORA $C800,Y") { addr = 0xc800.toShort; yr = 0x02 } {
           context("$C802 = $00") { setupOpArg(addr, yr, 0x00) } {
-            it("meets preconditions") { assert(memoryRead(0xc802) == 0x00) }
-            it { expect { operation }.notToChange { AC } }
-            it { expect { operation }.notToChange { ZF } }
-            it { expect { operation }.notToChange { SF } }
+            it("meets preconditions") { assert(memoryRead(0xc802)() == 0x00) }
+            it { expect2 { operation }.notToChange { AC } }
+            it { expect2 { operation }.notToChange { ZF } }
+            it { expect2 { operation }.notToChange { SF } }
           }
 
           context("$C802 = $01") { setupOpArg(addr, yr, 0x01) } {
-            it("meets preconditions") { assert(memoryRead(0xc802) == 0x01) }
-            it { expect { operation }.toChange { AC }.from(0x00).to(0x01) }
-            it { expect { operation }.toChange { ZF }.from(true).to(false) }
-            it { expect { operation }.notToChange { SF } }
+            it("meets preconditions") { assert(memoryRead(0xc802)() == 0x01) }
+            it { expect2 { operation }.toChange { AC }.from(0x00).to(0x01) }
+            it { expect2 { operation }.toChange { ZF }.from(true).to(false) }
+            it { expect2 { operation }.notToChange { SF } }
           }
 
           context("$C802 = $80") { setupOpArg(addr, yr, 0x80) } {
-            it("meets preconditions") { assert(memoryRead(0xc802) == 0x80) }
-            it { expect { operation }.toChange { AC }.from(0x00).to(0x80) }
-            it { expect { operation }.toChange { ZF }.from(true).to(false) }
-            it { expect { operation }.toChange { SF }.from(false).to(true) }
+            it("meets preconditions") { assert(memoryRead(0xc802)() == 0x80) }
+            it { expect2 { operation }.toChange { AC }.from(0x00).to(0x80) }
+            it { expect2 { operation }.toChange { ZF }.from(true).to(false) }
+            it { expect2 { operation }.toChange { SF }.from(false).to(true) }
           }
 
           context("$C802 = $ff") { setupOpArg(addr, yr, 0xff) } {
-            it("meets preconditions") { assert(memoryRead(0xc802) == 0xff) }
-            it { expect { operation }.toChange { AC }.from(0x00).to(0xff) }
-            it { expect { operation }.toChange { ZF }.from(true).to(false) }
-            it { expect { operation }.toChange { SF }.from(false).to(true) }
+            it("meets preconditions") { assert(memoryRead(0xc802)() == 0xff) }
+            it { expect2 { operation }.toChange { AC }.from(0x00).to(0xff) }
+            it { expect2 { operation }.toChange { ZF }.from(true).to(false) }
+            it { expect2 { operation }.toChange { SF }.from(false).to(true) }
           }
         }
       }
@@ -50,9 +50,9 @@ class AbsoluteYSpec extends FunOperationsSpec {
         context("AC = $00; YR = $02") { AC = 0x00; YR = 0x02 } {
           context("ORA $C8FF,Y") { addr = 0xc8ff.toShort; yr = 0x02 } {
             context("$C901 = $01") { setupOpArg(addr, yr, 0x01) } {
-              it("meets preconditions") { assert(memoryRead(0xc901) == 0x01) }
-              it { expect { operation }.toChange { AC }.from(0x00).to(0x01) }
-              it { expect { operation }.toUseCycles(0x05) }
+              it("meets preconditions") { assert(memoryRead(0xc901)() == 0x01) }
+              it { expect2 { operation }.toChange { AC }.from(0x00).to(0x01) }
+              it { expect2 { operation }.toUseCycles(0x05) }
             }
           }
         }
@@ -60,37 +60,37 @@ class AbsoluteYSpec extends FunOperationsSpec {
     }
 
     testOpCode(OpCode_AND_ABSY) {
-      it { expect { operation }.toAdvancePC(0x03) }
-      it { expect { operation }.toUseCycles(0x04) }
+      it { expect2 { operation }.toAdvancePC(0x03) }
+      it { expect2 { operation }.toUseCycles(0x04) }
 
       context("AC = $ff; YR = $02") { AC = 0xff; YR = 0x02; ZF = false; SF = true } {
         context("AND $C800,Y") { addr = 0xc800.toShort; yr = 0x02 } {
           context("$C802 = $00") { setupOpArg(addr, yr, 0x00) } {
-            it("meets preconditions") { assert(memoryRead(0xc802) == 0x00) }
-            it { expect { operation }.toChange { AC }.from(0xff).to(0x00) }
-            it { expect { operation }.toChange { ZF }.from(false).to(true) }
-            it { expect { operation }.toChange { SF }.from(true).to(false) }
+            it("meets preconditions") { assert(memoryRead(0xc802)() == 0x00) }
+            it { expect2 { operation }.toChange { AC }.from(0xff).to(0x00) }
+            it { expect2 { operation }.toChange { ZF }.from(false).to(true) }
+            it { expect2 { operation }.toChange { SF }.from(true).to(false) }
           }
 
           context("$C802 = $01") { setupOpArg(addr, yr, 0x01) } {
-            it("meets preconditions") { assert(memoryRead(0xc802) == 0x01) }
-            it { expect { operation }.toChange { AC }.from(0xff).to(0x01) }
-            it { expect { operation }.notToChange { ZF } }
-            it { expect { operation }.toChange { SF }.from(true).to(false) }
+            it("meets preconditions") { assert(memoryRead(0xc802)() == 0x01) }
+            it { expect2 { operation }.toChange { AC }.from(0xff).to(0x01) }
+            it { expect2 { operation }.notToChange { ZF } }
+            it { expect2 { operation }.toChange { SF }.from(true).to(false) }
           }
 
           context("$C802 = $80") { setupOpArg(addr, yr, 0x80) } {
-            it("meets preconditions") { assert(memoryRead(0xc802) == 0x80) }
-            it { expect { operation }.toChange { AC }.from(0xff).to(0x80) }
-            it { expect { operation }.notToChange { ZF } }
-            it { expect { operation }.notToChange { SF } }
+            it("meets preconditions") { assert(memoryRead(0xc802)() == 0x80) }
+            it { expect2 { operation }.toChange { AC }.from(0xff).to(0x80) }
+            it { expect2 { operation }.notToChange { ZF } }
+            it { expect2 { operation }.notToChange { SF } }
           }
 
           context("$C802 = $ff") { setupOpArg(addr, yr, 0xff) } {
-            it("meets preconditions") { assert(memoryRead(0xc802) == 0xff) }
-            it { expect { operation }.notToChange { AC } }
-            it { expect { operation }.notToChange { ZF } }
-            it { expect { operation }.notToChange { SF } }
+            it("meets preconditions") { assert(memoryRead(0xc802)() == 0xff) }
+            it { expect2 { operation }.notToChange { AC } }
+            it { expect2 { operation }.notToChange { ZF } }
+            it { expect2 { operation }.notToChange { SF } }
           }
         }
       }
@@ -99,9 +99,9 @@ class AbsoluteYSpec extends FunOperationsSpec {
         context("AC = $ff; YR = $02") { AC = 0xff; YR = 0x02 } {
           context("AND $C8FF,Y") { addr = 0xc8ff.toShort; yr = 0x02 } {
             context("$C901 = $01") { setupOpArg(addr, yr, 0x01) } {
-              it("meets preconditions") { assert(memoryRead(0xc901) == 0x01) }
-              it { expect { operation }.toChange { AC }.from(0xff).to(0x01) }
-              it { expect { operation }.toUseCycles(0x05) }
+              it("meets preconditions") { assert(memoryRead(0xc901)() == 0x01) }
+              it { expect2 { operation }.toChange { AC }.from(0xff).to(0x01) }
+              it { expect2 { operation }.toUseCycles(0x05) }
             }
           }
         }
@@ -109,37 +109,37 @@ class AbsoluteYSpec extends FunOperationsSpec {
     }
 
     testOpCode(OpCode_EOR_ABSY) {
-      it { expect { operation }.toAdvancePC(0x03) }
-      it { expect { operation }.toUseCycles(0x04) }
+      it { expect2 { operation }.toAdvancePC(0x03) }
+      it { expect2 { operation }.toUseCycles(0x04) }
 
       context("AC = $80; YR = $02") { AC = 0x80; YR = 0x02; ZF = false; SF = true } {
         context("EOR $C800,Y") { addr = 0xc800.toShort; yr = 0x02 } {
           context("$C802 = $00") { setupOpArg(addr, yr, 0x00) } {
-            it("meets preconditions") { assert(memoryRead(0xc802) == 0x00) }
-            it { expect { operation }.notToChange { AC } }
-            it { expect { operation }.notToChange { ZF } }
-            it { expect { operation }.notToChange { SF } }
+            it("meets preconditions") { assert(memoryRead(0xc802)() == 0x00) }
+            it { expect2 { operation }.notToChange { AC } }
+            it { expect2 { operation }.notToChange { ZF } }
+            it { expect2 { operation }.notToChange { SF } }
           }
 
           context("$C802 = $01") { setupOpArg(addr, yr, 0x01) } {
-            it("meets preconditions") { assert(memoryRead(0xc802) == 0x01) }
-            it { expect { operation }.toChange { AC }.from(0x80).to(0x81) }
-            it { expect { operation }.notToChange { ZF } }
-            it { expect { operation }.notToChange { SF } }
+            it("meets preconditions") { assert(memoryRead(0xc802)() == 0x01) }
+            it { expect2 { operation }.toChange { AC }.from(0x80).to(0x81) }
+            it { expect2 { operation }.notToChange { ZF } }
+            it { expect2 { operation }.notToChange { SF } }
           }
 
           context("$C802 = $80") { setupOpArg(addr, yr, 0x80) } {
-            it("meets preconditions") { assert(memoryRead(0xc802) == 0x80) }
-            it { expect { operation }.toChange { AC }.from(0x80).to(0x00) }
-            it { expect { operation }.toChange { ZF }.from(false).to(true) }
-            it { expect { operation }.toChange { SF }.from(true).to(false) }
+            it("meets preconditions") { assert(memoryRead(0xc802)() == 0x80) }
+            it { expect2 { operation }.toChange { AC }.from(0x80).to(0x00) }
+            it { expect2 { operation }.toChange { ZF }.from(false).to(true) }
+            it { expect2 { operation }.toChange { SF }.from(true).to(false) }
           }
 
           context("$C802 = $ff") { setupOpArg(addr, yr, 0xff) } {
-            it("meets preconditions") { assert(memoryRead(0xc802) == 0xff) }
-            it { expect { operation }.toChange { AC }.from(0x80).to(0x7f) }
-            it { expect { operation }.notToChange { ZF } }
-            it { expect { operation }.toChange { SF }.from(true).to(false) }
+            it("meets preconditions") { assert(memoryRead(0xc802)() == 0xff) }
+            it { expect2 { operation }.toChange { AC }.from(0x80).to(0x7f) }
+            it { expect2 { operation }.notToChange { ZF } }
+            it { expect2 { operation }.toChange { SF }.from(true).to(false) }
           }
         }
       }
@@ -148,9 +148,9 @@ class AbsoluteYSpec extends FunOperationsSpec {
         context("AC = $80; YR = $02") { AC = 0x80; YR = 0x02 } {
           context("EOR $C8FF,Y") { addr = 0xc8ff.toShort; yr = 0x02 } {
             context("$C901 = $01") { setupOpArg(addr, yr, 0x01) } {
-              it("meets preconditions") { assert(memoryRead(0xc901) == 0x01) }
-              it { expect { operation }.toChange { AC }.from(0x80).to(0x81) }
-              it { expect { operation }.toUseCycles(0x05) }
+              it("meets preconditions") { assert(memoryRead(0xc901)() == 0x01) }
+              it { expect2 { operation }.toChange { AC }.from(0x80).to(0x81) }
+              it { expect2 { operation }.toUseCycles(0x05) }
             }
           }
         }